Fix restoring PB and PC when an exception is thrown.
When an exception is thrown, a pointer to the Data::s_exceptionInstructions array is returned. This is then converted into a negative offset relative to PB (which points to the start of the CodeBlock::m_instructions array). On CHERI, when this negative offset is added to PB, a CHERI length violation occurs. To fix this, in the case of an exception, we set PB to be the address of Data::s_exceptionInstructions and PC to 0.

Also comment out unnecessary sign-extension in _llint_op_get_by_val