Fix integer compare and the getPropery() macro.
The offset of the property to load is an unsigned int. We revert the asm
instruction to load an integer and also ensure we store an int when setting up
the instructions array (in CodeBlock::finishCreation) for global and unresolved
properties. The bug occurred in the op_get_from_scope operation, because we
were storing an intcap_t in instructions[i+6] in CodeBlock::finishCreation, but
this was being overriten with an integer in tryCacheGetFromScopeGlobal
(CommonSlowPaths.h).