Fix intialising of the offset operand for put_to_scope.
The value being stored for this operand is an intptr_t and previously it was
cast to a pointer before being stored in instructions[i+6], regardless of
whether it was an unsigned int or a pointer. However, the asm would always read
it as a 32-bit integer. This meant that we were always reading back 0 in the
putProperty() macro.

I have replicated the fix from get_from_scope here because it's the same
problem.