Fix stack-alignment-related constants
StackAlignmentSlots corresponds to how many slots the stack alignment
corresponds to. We previously had a value of 2 when it should be 1 (we align as
16-bytes or 32-bytes on CHERI-128 and CHERI-256 respectively, and the slot size
is the same - i.e. as it corresponds to the capability size). As a result of
the incorrect value, jsc was overwriting data on the stack with 0xa (meaning
undefined). This manifested when the m_vm member field of the VMEntryRecord
object was 0xa.